# spatial conv layer for Xception