From discussion in #778 : when group == num_output (spatial conv), the memory consuming is unbearable. This PR implements spaital conv in conv_layer.cu. This implementation could be used Xception, a prototxt can be found here: https://github.com/yihui-he/Xception-caffe/blob/master/xception-50.prototxt . 
An example of usage:
```
layer {
  name: "interstellar2a_branch1_s"
  type: "Convolution"
  bottom: "pool1"
  top: "interstellar2a_branch1_s"
  param {
    lr_mult: 1
    decay_mult: 1
  }
  convolution_param {
    num_output: 64
    bias_term: false
    pad: 1
    kernel_size: 3
    group: 64
    stride: 1
    weight_filler {
      type: "msra"
    }
    engine: CAFFE
  }
}
```