@Noiredd 2018-02-21
I compared performance of this implementation with the latest master on my GTX TITAN Z, taking a 1x256x256x256 input and a conv layer with `num_output: 256`, `group: 256`, `kernel_size: 5`, `stride: 1`, `pad: 1`. Below I show a forward-backward iteration time and VRAM allocated as shown by `nvidia-smi`:

implementation | iteration time | VRAM taken
--- | --- | ---
master | 38.7 ms | 324 MB
this PR | 20.2 ms | 324 MB

For a 1x32x256x256 input and a conv with `num_output: 32`, `group: 32`, `kernel_size: 5`, `stride: 1`, `pad: 1`:

implementation | iteration time | VRAM taken
--- | --- | ---
master | 297.5 ms | 1924 MB
this PR | 143.5 ms | 1924 MB

So it was faster, but I was unable to observe any RAM savings.

---

Additional issues with this PR:
* it does not implement biases (though there is some commented-out code for a forward pass that could be easily made to work), failing the forward tests (`ConvolutionLayerTest/2.TestSimpleConvolutionGroup,`),
* even with this fixed, it fails the backward tests (`ConvolutionLayerTest/2.TestGradientGroup`),
* from what I see, it does not implement dilation (though we do not seem to have a test for group+dilation),
* numerous lint issues.

Are you willing to continue working on this, @yihui-he? RAM savings and speedups for grouped conv would be **very** useful, but this needs some more attention.