@naibaf7 2018-02-21
There is also an implementation that does not consume any extra memory for convolution here:
https://github.com/naibaf7/caffe/tree/master/src/caffe/libdnn (CUDA and OpenCL)
and here:
https://github.com/BVLC/caffe/blob/opencl/src/caffe/greentea/libdnn_conv.cpp (CUDA and OpenCL)
and here:
https://github.com/BVLC/caffe/blob/opencl/src/caffe/greentea/libdnn_conv_spatial.cpp (OpenCL)